CrimsonSkyline.jl

trace

To a large extent, Nodes have local control over the behavior of inference algorithms via their interpretation. There are a variety of Interpretations. The type hierarchy is flat:

abstract type Interpretation end
struct Nonstandard <: Interpretation end
struct Standard <: Interpretation end
struct Replayed <: Interpretation end
struct Conditioned <: Interpretation end 
struct Blocked <: Interpretation end
struct Deterministic <: Interpretation end
struct Proposed <: Interpretation end

CrimsonSkyline.ParametricNode — Type

mutable struct ParametricNode{A, D, T} <: Node
    address :: A
    dist :: D
    value :: Maybe{T}
    logprob :: Float64
    logprob_sum :: Float64
    observed :: Bool
    pa :: Array{Node, 1}
    ch :: Array{Node, 1}
    interpretation :: Union{Interpretation, Vector{Interpretation}}
    last_interpretation :: Union{Interpretation, Vector{Interpretation}}
end

A Node that can be used with arbitrary code for which rand and Distributionss.logpdf are defined.

CrimsonSkyline.SampleableNode — Type

mutable struct SampleableNode{A, T} <: Node
    address :: A
    dist :: Sampleable
    value :: Maybe{T}
    logprob :: Float64
    logprob_sum :: Float64
    observed :: Bool
    pa :: Array{Node, 1}
    ch :: Array{Node, 1}
    interpretation :: Union{Interpretation, Vector{Interpretation}}
    last_interpretation :: Union{Interpretation, Vector{Interpretation}}
end

A Node that is restricted to be used with any Sampleable from Distributions.jl.

CrimsonSkyline.Trace — Type

abstract type Trace end

Base type for all traces. Traces support the following Base methods: setindex!, getindex, keys, values, and length.

CrimsonSkyline.TypedTrace — Type

mutable struct TypedTrace{A, T} <: Trace
    trace :: OrderedDict{A, SampleableNode{A, T}}
    logprob_sum :: Float64
end

Trace that can hold nodes of the specific address (A) and value (T) types.

CrimsonSkyline.UntypedTrace — Type

mutable struct UntypedTrace
    trace :: OrderedDict{Any, Node}
    logprob_sum :: Float64
end

Trace that can hold nodes with all address and value types.

CrimsonSkyline.aic — Method

function aic(t :: Trace)

Computes the Akaike Information Criterion for a single trace (thus replacing the definition) with "maximum likelihood" by one with "likelihood". The formula is

\[\text{AIC}(t)/2 = |\text{params}(t)| - \ell(t),\]

where $\text{params}(t)|$ is the number of non-observed and non-deterministic sample nodes.

CrimsonSkyline.connect_pa_ch! — Method

function connect_pa_ch!(t :: Trace, pa, a)

Connects parent and child nodes. Adds child nodes to parent's ch and parent nodes to child's pa.

CrimsonSkyline.input — Method

input(t :: Trace, a, d)

Track a model input. Used only in graph intermediate representation and factor graph.

CrimsonSkyline.interpret_latent! — Method

function interpret_latent!(t :: Trace, i :: Interpretation)

Changes the interpretation of all latent nodes in t to have interpretation == i

CrimsonSkyline.loglikelihood — Method

function loglikelihood(t :: Trace)

Computes and returns the log likelihood of the observed data under the model:

\[\ell(t) = \sum_{a:\ [a \in \text{keys}(t)] \wedge [\text{interpretation}(a) = \text{Standard}]} \log p(t[a])\]

CrimsonSkyline.logprob! — Method

function logprob!(t :: Trace)

Computes the joint log probability to the trace and assigns it to t.logprob_sum.

CrimsonSkyline.logprob — Method

function logprob(t :: Trace)

Computes and returns the joint log probability of the trace:

\[\log p(t) = \sum_{a \in \text{keys}(t)}\log p(t[a])\]

CrimsonSkyline.node — Method

function node(value, address :: A, dist :: D, is_obs :: Bool, i :: Interpretation) where {A, D}

Outer constructor for Node where data is passed during construction. Data type is inferred from the passed data.

CrimsonSkyline.node — Method

function node(value, address :: A, dist :: D, is_obs :: Bool, i :: Interpretation) where {A, D}

Outer constructor for Node where data is passed during construction. Data type is inferred from the passed data.

CrimsonSkyline.node — Method

function node(T :: DataType, address :: A, dist :: D, is_obs :: Bool, i :: Interpretation) where {A, D}

Outer constructor for Node where no data is passed during construction.

CrimsonSkyline.node — Method

function node(T :: DataType, address :: A, dist :: D, is_obs :: Bool, i :: Interpretation) where {A, D}

Outer constructor for Node where no data is passed during construction.

CrimsonSkyline.observe — Method

function observe(t :: Trace, a, d, s; pa = ())

If d is not nothing an alias for calling sample with standard interpretation. Otherwise, an alias for calling sample with nonstandard interpretation.

CrimsonSkyline.pa_from_trace — Method

function pa_from_trace(t :: Trace, pa)

Collects nodes in trace corresponding to an iterable of parent addresses pa.

CrimsonSkyline.plate — Method

function plate(t::Trace, op::F, a, d, s::Int64, i::Conditioned; pa = ()) where F<:Function

Plate over conditioned variables.

CrimsonSkyline.plate — Method

function plate(t::Trace, op::F, a, d, s::Int64, i::Blocked; pa = ()) where F<:Function

Plate over blocked variables.

CrimsonSkyline.plate — Method

function plate(t::Trace, op::F, a, d, s::Int64, i::Nonstandard; pa = ()) where F<:Function

Plate over latent variables.

CrimsonSkyline.plate — Method

function plate(t::Trace, op::F, a, d, s::Int64, i::Replayed; pa = ()) where F<:Function

Plate over replayed variables. Note that this method assumes and does not check that the value to be replayed v satisfies length(v) == s.

CrimsonSkyline.plate — Method

function plate(t::Trace, op::F, a, d, s::Int64; pa = ()) where F<:Function

Sample or observe a vector of random variables at a single site instead of multiple. This can speed up inference since the number of sites in the model will no longer scale with dataset size (though numerical value computation is still linear in dataset size).

Example usage: instead of

vals = [sample(t, "val $i", Geometric()) for i in 1:N]

we can write

vals = plate(t, sample, "val", Geometric(), N)

Mathematically, this is equivalent to the product $p(z) = \prod_n p(z_n)$ and treating it as the single object $p(z)$ instead of the $N$ objects $p(z_n)$.

CrimsonSkyline.plate — Method

function plate(t::Trace, op::F, a, d, v::Vector{T}; pa = ()) where {T, F<:Function}

Plate over observed variables, i.e., a plated component of model likelihood. v is the vector of observations, while op is likely observe.

Example usage: instead of

for (i, d) in enumerate(data)
    observe(t, "data $i", Normal(loc, scale), d)
end

we can write

plate(t, observe, "data", Normal(loc, scale), data)

CrimsonSkyline.propose — Method

propose(t :: Trace, a, d)

Propose a value for the address a in trace t from the distribution d.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, params, i :: Blocked; pa = ())

Samples from d passing the optional arguments params, deletes the node stored at address a from trace t, and returns the sampled value.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, f, v, i :: Deterministic; pa = ())

Creates a deterministic node mapping the tuple of data v through function f, storing the value in trace t at address a.

Infers input type from v

Maps tuple of data v through function f, yielding r = f(v...)

Creates a deterministic node and stores it in t at address a

Optionally adds nodes corresponding to the addresses in pa as parent nodes

Returns r

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, params, i :: Nonstandard; pa = ())

Samples from distribution d into trace t at address a.

Samples a value from d passing the optional arguments params

Creates a sample node

Adds the sample node to trace t at value a

Optionally adds nodes corresponding to the addresses in pa as parent nodes

Returns the sampled value

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, params, i :: Replayed; pa = ())

Replays the sampled node through the trace.

If a is not in t's address set, calls sample(t, a, d, NONSTANDARD; pa = pa).

Creates a sample node that copies the value from the last node stored in the trace at address a.

Adds the sample node to trace t at value a

Optionally adds nodes corresponding to the addresses in pa as parent nodes

Resets the node's interpretation to the original interpretation

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, s, i :: Standard; pa = ())

Scores an observed value s against the distribution d, storing the value in trace t at address a and optionally adds nodes corresponding to the addresses in pa as parent nodes.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, params, ii :: Array{Interpretation, 1}; pa = ())

Sequentially apply sample statements with interpretations as given in ii. This is used to depth-first traverse the interpretation graph.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, params; pa = ())

If a is in the set of trace addresses, calls sample using t[a]'s interpretation. Otherwise, calls sample using nonstandard interpretation.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, i :: Blocked; pa = ())

Samples from d, deletes the node stored at address a from trace t, and returns the sampled value.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, i :: Nonstandard; pa = ())

Samples from distribution d into trace t at address a.

Samples a value from d

Creates a sample node

Adds the sample node to trace t at value a

Optionally adds nodes corresponding to the addresses in pa as parent nodes

Returns the sampled value

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, i :: Replayed; pa = ())

Replays the sampled node through the trace.

If a is not in t's address set, calls sample(t, a, d, NONSTANDARD; pa = pa).

Creates a sample node that copies the value from the last node stored in the trace at address a.

Adds the sample node to trace t at value a

Optionally adds nodes corresponding to the addresses in pa as parent nodes

Resets the node's interpretation to the original interpretation

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, i :: Union{Standard,Conditioned}; pa = ())

Scores an observed value against the distribution d, storing the value in trace t at address a and optionally adds nodes corresponding to the addresses in pa as parent nodes.

This method is used by the condition effect. It will probably not be used by most users.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d, ii :: Array{Interpretation, 1}; pa = ())

Sequentially apply sample statements with interpretations as given in ii. This is used to depth-first traverse the interpretation graph.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, d; pa = ())

If a is in the set of trace addresses, calls sample using t[a]'s interpretation. Otherwise, calls sample using nonstandard interpretation.

CrimsonSkyline.trace — Method

trace(A, T)

This is the recommended way to construct a new typed trace. A is the address type, T is the value type.

CrimsonSkyline.trace — Method

trace()

This is the recommended way to construct a new trace.

CrimsonSkyline.transform — Method

function transform(t :: Trace, a, f :: F, v; pa = ()) where F <: Function

Alias for sample(t, a, f, v, DETERMINISTIC; pa = pa).

io

Saving and loading traces and SamplingResults is possible using the save and load functions. This functionality depends on JuliaDB.

Examples:

# simple example model
function io_model(t::Trace, dim::Int64, data::Int64)
    z = sample(t, "z", Dirichlet(ones(dim)))
    observe(t, "x", Categorical(z), data)
end

# saving a trace
t = trace()
dim = 10
data = 7
io_model(t, dim, data)
testpath = joinpath(@__DIR__, "TESTIO")
db_file = joinpath(testpath, "test.jdb")
save(t, db_file)

# loading a trace
loaded_trace = load(db_file)

# creating and saving some results
results = mh(io_model; params = (dim, data), burn = 0, thin = 1, num_iterations=10)
results_file = joinpath(testpath, "io_model.csm")
save(results, results_file)

# loading saved results and use to serve model
loaded_results = load(joinpath(testpath, "io_model.csm"))
updated_model = update(io_model, loaded_results)  # update effect, see effects.jl
new_t, _ = updated_model(trace(), dim, data)

CrimsonSkyline.load — Method

function load(f)

Load a Trace or SamplingResults object from file. The file must be saved in JuliaDB format with ending .jdb, which will be interpreted as a single saved trace, or must be a directory with ending .csm, which will be interpreted as a SamplingResults object.

CrimsonSkyline.load_csm — Method

function load_csm(f) :: SamplingResults

Loads a SamplingResults from a directory. The directory contains metadata.txt, which currently stores the interpretation of the SamplingResults (i.e., what kind of algorithm generated those results), and a file results.jdb, which is a JuliaDB table of the results.

CrimsonSkyline.load_jdb — Method

function load_jdb(f) :: Trace

Loads a serialized juliadb table from file f and converts it into a trace.

CrimsonSkyline.save — Method

function save(r :: SamplingResults, f)

Saves a SamplingResults to disk in the directory f.

CrimsonSkyline.save — Method

function save(t :: Trace, f)

Saves a trace to disk at the filepath f.

CrimsonSkyline.to_table — Method

function to_table(t :: Trace)

Turns a trace into a juliadb table. Does not store parent / child relationships.

effects

A library of functions that change the interpretation of some or all nodes in a trace.

CrimsonSkyline.block — Method

block(f :: F, t :: Trace) where F <: Function

Converts all traced randomness into untraced randomness.

CrimsonSkyline.block — Method

function block(f, t :: T, addresses) where T <: Trace

Given a stochastic function f, a trace t, and an iterable of addresses, converts traced randomness into untraced randomness.

Returns a tuple (t_new, g), where t_new is a trace and g is a function. The function signature of g is the same as that of f with the first argument removed; that is, if f(t :: Trace, params...), then g(params...). Computation is delayed, so each of the latent nodes in t_new has interpretation = BLOCKED. Calling g(params...) executes the computation and each latent node in t_new with an address in addresses is removed.

CrimsonSkyline.condition — Method

function condition(f, evidence :: Dict)

Condition a trace modified by f on evidence, which maps addresses to observed evidence associated with that address. Returns a function with call signature identical to that of f and return signature (t :: Trace, rtype) where rtype is the return type of f.

CrimsonSkyline.replace — Method

function replace(f, r :: Dict)

Given a mapping r from addresses to distribution-like (currently Distributions or Array{Any, 1}s), replaces the current distributions at that set of addresses with this set of distributions. Returns a function g that has return signature (t :: Trace, rval) where rval is a return value of f.

CrimsonSkyline.replace — Method

function replace(t :: Trace, r :: Dict)

Given a mapping r from addresses to distribution-like (currently Distributions or Array{Any, 1}s), replaces the current distributions at that set of addresses with this set of distributions. Returns the modified trace.

CrimsonSkyline.replay — Method

function replay(f, t :: T) where T <: Trace

Given a stochastic function f and a trace t, makes sample calls behave as though they had sampled the values in t at the corresponding addresses.

Returns a tuple (t_new, g), where t_new is a trace and g is a function. The function signature of g is the same as that of f with the first argument removed; that is, if f(t :: Trace, params...), then g(params...). Computation is delayed, so each of the latent nodes in t_new has interpretation = REPLAYED. Calling g(params...) executes the computation and each latent node in t_new reverts to its original interpretation.

CrimsonSkyline.rewrite — Method

function rewrite(f, t :: T, r :: Dict) where T <: Trace

Rewrites the history of the trace to make it appear as if the values in the trace were sampled at the addresses in the keys of r from the corresponding distributions in the values of r. Returns a function with call signature g(params...) that returns (t :: Trace, rval), where rval is the return type of f.

CrimsonSkyline.update — Method

function update(f, r :: SamplingResults{I}) where I <: InferenceType

Given a stochastic function f and a SamplingResults r, update the prior predictive to the posterior predictive by jointly replacing all latent sample sites with the joint empirical posterior. Returns a stochastic function g with the same call signature as f. This function will modify in place the trace passed into it as the first argument.

basic sampling methods

Simple samplers such as forward and rejection sampling can be done "by hand", but convenience methods are implemented to facilitate postprocessing of results.

CrimsonSkyline.forward_sampling — Method

function forward_sampling(f; params = (), num_iterations = 1)

Draws samples from the model's joint density. Equivalent to calling f in a loop num_iterations times, but results are collected in a NonparametricSamplingResults for easier postprocessing.

CrimsonSkyline.rejection — Method

function rejection(f, log_l :: Float64, params...)

Samples from the prior with the hard likelihood constraint $\log L_k >$ log_l.

Args:

f stochastic function. Must have signature f(t :: Trace, params...)

log_l: current log likelihood threshold

params: any additional arguments to pass to f

importance

Importance sampling algorithms and utilities. Currently the following importance sampling algorithms are implemented:

Likelihood weighting

Generic user-defined proposal

CrimsonSkyline.importance_sampling — Method

function importance_sampling(f, q, types::Tuple{DataType,DataType}; params = (), nsamples :: Int = 1)

Given a stochastic function f, a proposal function q, and a tuple of params to pass to f and q, compute nsamples iterations of importance sampling. q must have the same input signature as f. Returns a SamplingResults instance.

CrimsonSkyline.importance_sampling — Method

function importance_sampling(f, q; params = (), nsamples :: Int = 1)

Given a stochastic function f, a proposal function q, and a tuple of params to pass to f and q, compute nsamples iterations of importance sampling. q must have the same input signature as f. Returns a SamplingResults instance.

CrimsonSkyline.is_step — Method

function is_step(f, q, types::Tuple{DataType,DataType}; params = ())

Perform one step of importance sampling – draw a single sample from the proposal q, replay it through f, and record the log weight as $\log W_n = \log p(x, z_n) - \log q(z_n)$. Returns a tuple (log weight, rval, trace).

CrimsonSkyline.is_step — Method

function is_step(f, q; params = ())

Perform one step of importance sampling – draw a single sample from the proposal q, replay it through f, and record the log weight as $\log W_n = \log p(x, z_n) - \log q(z_n)$. Returns a tuple (log weight, rval, trace).

CrimsonSkyline.likelihood_weighting — Method

function likelihood_weighting(f, types::Tuple{DataType, DataType}, params...; nsamples :: Int = 1)

Given a stochastic function f and arguments to the function params..., executes nsamples iterations of importance sampling by using the prior as a proposal distribution. The importance weights are given by $\log W_n = \ell(t_n)$. Returns an SamplingResults instance.

CrimsonSkyline.likelihood_weighting — Method

function likelihood_weighting(f, params...; nsamples :: Int = 1)

Given a stochastic function f and arguments to the function params..., executes nsamples iterations of importance sampling by using the prior as a proposal distribution. The importance weights are given by $\log W_n = \ell(t_n)$. Returns an SamplingResults instance.

CrimsonSkyline.likelihood_weighting_results — Method

likelihood_weighting_results()

Outer constructor for SamplingResults.

CrimsonSkyline.log_evidence — Method

function log_evidence(r :: SamplingResults{LikelihoodWeighting})

Computes the log evidence (log partition function),

$\log Z \equiv \log p(x) \approx -\log N_{\text{samples}} + \log \sum_{n=1}^{N_{\text{samples}}} W_n.$

CrimsonSkyline.lw_step — Method

function lw_step(f, types::Tuple{DataType, DataType}, params...)

Perform one step of likelihood weighting – draw a single proposal from the prior and compute the log weight as equal to the likelihood. Returns a tuple (log weight, rval, trace).

CrimsonSkyline.lw_step — Method

function lw_step(f, params...)

Perform one step of likelihood weighting – draw a single proposal from the prior and compute the log weight as equal to the likelihood. Returns a tuple (log weight, rval, trace).

CrimsonSkyline.normalized_weights — Method

function normalized_weights(r :: SamplingResults{LikelihoodWeighting})

Computes the normalized weights $w_n$ from unnormalized weights $W_n$:

$w_n = W_n / p(x) = \exp\{\ell(t_n) - \log Z\}.$

CrimsonSkyline.sample — Method

function sample(r :: SamplingResults{LikelihoodWeighting}, k, n :: Int)

Draws n samples from the empirical marginal posterior at address k.

metropolis

Metropolis algorithm and utilities. Currently the following algorithms are implemented for both FOPPL and HOPPL programs:

Independent prior proposal

Arbitrary single- or multi-site proposal

Here are two examples of inference using the arbitrary MH step interface. Consider the following generative function:

function normal_model(t :: Trace, data :: Vector{Float64})
    loc = sample(t, :loc, Normal(0.0, 10.0))
    scale = sample(t, :scale, LogNormal())
    for i in 1:length(data)
        observe(t, (:obs, i), Normal(loc, scale), data[i])
    end
end

To learn an approximate posterior for :loc and :scale, we will introduce two proposal kernels:

loc_proposal(old_t :: Trace, new_t :: Trace, data) = propose(new_t, :loc, Normal(old_t[:loc].value, 0.25))
scale_proposal(old_t :: Trace, new_t :: Trace, data) = propose(new_t, :scale, truncated(Normal(old_t[:scale].value, 0.25), 0.0, Inf))

Note that while loc_proposal is symmetric, scale_proposal is not. To perform inference, we pass these kernels to mh_step in a loop after first drawing a random trace:

t = trace()
normal_model(t, data)
for i in 1:niter
    t = mh_step(t, normal_model, loc_proposal; params = (data,))
    t = mh_step(t, normal_model, scale_proposal; params = (data,))
end

In this case, inference was fairly successful:

[ Info: True loc = 4.0
[ Info: True std = 1.0
[ Info: inferred E[loc] = 4.022688081613082
[ Info: inferred E[scale] = 0.9696559373495869
[ Info: approximate p(x) = sum_z p(x|z) = -138.63530736205144

As a less trivial (but still contrived!) inference example, we can infer the posterior distribution of a latent discrete random variable in an open-universe model:

function random_sum_model(t :: Trace, data)
    n = sample(t, :n, Geometric(0.1))
    loc = 0.0
    for i in 1:(n + 1)
        loc += sample(t, (:loc, i), Normal())
    end
    obs = Array{Float64, 1}()
    for j in 1:length(data)
        o = observe(t, (:data, j), Normal(loc, 1.0), data[j])
        push!(obs, o)
    end
    obs
end

(N.B.: we write the model is this form for pedagogic reasons; there is a far more efficient way to express the latent structure of this model, namely $n \sim \text{Geometric}(0.1)$, $\text{loc} \sim \text{Normal}(0, n)$.) We are interested in learning the posterior distribution of :n. We introduce two proposal distributions, one for the latent discrete rv and another generic proposal for the location increments:

function random_n_proposal(old_trace, new_trace, params...)
    old_n = float(old_trace[:n].value)
    if old_n > 0
        propose(new_trace, :n, Poisson(old_n))
    else
        propose(new_trace, :n, Poisson(1.0))
    end
end

gen_loc_proposal(old_trace, new_trace, ix, params...) = propose(new_trace, (:loc, ix), Normal(old_trace[(:loc, ix)].value, 0.25))

We again conduct inference by simply applying proposals within a loop. This time, the number of location increment proposals we need to construct is dependent on the sampled values of the latent random variable. We can either create these proposals on the fly as they're needed or create what is nearly guaranteed to be enough of them before any inference is performed, e.g., loc_proposals = [(o, n, params...) -> gen_loc_proposal(o, n, i, params...) for i in 1:100]. Now that we have what we need, we can conduct inference:

t = trace()
random_sum_model(t, data)
...
for i in 1:niter
    t = mh_step(t, random_sum_model, random_n_proposal; params=(data,))
    for j in 1:(t[:n].value + 1)
        t = mh_step(t, random_sum_model, loc_proposals[j]; params=(data,))
    end
    push!(ns, t[:n].value)
end

Our inference results look promising:

[ Info: True :n = 9
[ Info: Posterior E[:n] = 7.581604598850287

For more examples, check back soon.

CrimsonSkyline.accept — Method

function accept(t :: Trace, new_t :: Trace, log_a :: Float64)

Stochastic function that either returns new_t if accepted or returns t if not accepted.

CrimsonSkyline.copy_common! — Method

function copy_common!(old_t :: Trace, new_t :: Trace)

Copies nodes from old_t into new_t for all addresses in the intersection of their address sets.

CrimsonSkyline.log_acceptance_ratio — Method

function log_acceptance_ratio(t :: Trace, t_proposed :: Trace, p :: Prior)

Computes the log acceptance ratio of a Metropolis step when using the independent prior proposal algorithm:

$\log \alpha = \ell(t_{\text{proposed}}) - \ell(t_{\text{original}})$

CrimsonSkyline.loglatent — Method

function loglatent(t :: Trace)

Computes the joint log probability of all latent variables in a trace, $\log p(t) - \ell(t)$.

CrimsonSkyline.logprob — Method

function logprob(t0 :: Trace, t1 :: Trace)

Computes the proposal log probability $q(t_1 | t_0)$.

This expression has two parts: log probability that is generated at the proposed site(s), and log probability that is generated at the sites that are present in t1 but not in t0.

CrimsonSkyline.mh — Method

function mh(f, types::Tuple{DataType, DataType}; params = (), burn = 1000, thin = 50, num_iterations = 10000)

Generic Metropolis algorithm using draws from the prior.

Args:

f: stochastic function. Must have call signature f(t :: Trace, params...)

params: addditional arguments to pass to f and each of the proposal kernels.

burn: number of samples to discard at beginning of markov chain

thin: keep only every thin-th draw. E.g., if thin = 100, only every 100-th trace will be kept.

num_iterations: total number of steps to take in the markov chain

CrimsonSkyline.mh — Method

function mh(f; params = (), burn = 1000, thin = 50, num_iterations = 10000)

Generic Metropolis algorithm using draws from the prior.

Args:

f: stochastic function. Must have call signature f(t :: Trace, params...)

params: addditional arguments to pass to f and each of the proposal kernels.

burn: number of samples to discard at beginning of markov chain

thin: keep only every thin-th draw. E.g., if thin = 100, only every 100-th trace will be kept.

num_iterations: total number of steps to take in the markov chain

CrimsonSkyline.mh — Method

function mh(f, qs :: A, types::Tuple{DataType,DataType}; params = (), burn = 100, thin = 10, num_iterations = 10000, inverse_verbosity = 100) where A <: AbstractArray

Generic Metropolis algorithm using user-defined proposal kernels.

Args:

f: stochastic function. Must have call signature f(t :: Trace, params...)

qs: array-like of proposal kernels. Proposal kernels are applied sequentially in the order that they appear in this array. Proposal kernels must have the signature q(old_t :: Trace, new_t :: Trace, params...) where it must take in at least the same number of arguments in params as f.

params: addditional arguments to pass to f and each of the proposal kernels.

burn: number of samples to discard at beginning of markov chain

thin: keep only every thin-th draw. E.g., if thin = 100, only every 100-th trace will be kept.

num_iterations: total number of steps to take in the markov chain

inverse_verbosity: every inverse_verbosity iterations, a stattus report will be logged.

CrimsonSkyline.mh — Method

function mh(f, qs :: A; params = (), burn = 100, thin = 10, num_iterations = 10000, inverse_verbosity = 100) where A <: AbstractArray

Generic Metropolis algorithm using user-defined proposal kernels.

Args:

f: stochastic function. Must have call signature f(t :: Trace, params...)

qs: array-like of proposal kernels. Proposal kernels are applied sequentially in the order that they appear in this array. Proposal kernels must have the signature q(old_t :: Trace, new_t :: Trace, params...) where it must take in at least the same number of arguments in params as f.

params: addditional arguments to pass to f and each of the proposal kernels.

burn: number of samples to discard at beginning of markov chain

thin: keep only every thin-th draw. E.g., if thin = 100, only every 100-th trace will be kept.

num_iterations: total number of steps to take in the markov chain

inverse_verbosity: every inverse_verbosity iterations, a stattus report will be logged.

CrimsonSkyline.mh_step — Method

function mh_step(t :: Trace, f, q, types::Tuple{DataType,DataType}; params = (), return_val :: Bool = false)

A generic Metropolis step using an arbitrary proposal kernel.

Given a trace t, a stochastic function f with signature f(t :: Trace, params...) a stochastic function q with signature q(old_trace :: Trace, new_trace :: Trace, params...), generates a proposal from q and accepts based on the log acceptance probability:

$\log \alpha = \log p(t_{\text{new}}) - \log q(t_{\text{new}}|t_{\text{old}}) - [\log p(t_{\text{old}}) - \log q(t_{\text{old}} | t_{\text{new}})].$

CrimsonSkyline.mh_step — Method

function mh_step(t :: Trace, f, q; params = (), return_val :: Bool = false)

A generic Metropolis step using an arbitrary proposal kernel.

Given a trace t, a stochastic function f with signature f(t :: Trace, params...) a stochastic function q with signature q(old_trace :: Trace, new_trace :: Trace, params...), generates a proposal from q and accepts based on the log acceptance probability:

$\log \alpha = \log p(t_{\text{new}}) - \log q(t_{\text{new}}|t_{\text{old}}) - [\log p(t_{\text{old}}) - \log q(t_{\text{old}} | t_{\text{new}})].$

CrimsonSkyline.mh_step — Method

function mh_step(t :: Trace, f, types::Tuple{DataType, DataType}; params = (), return_val :: Bool = false)

An independent prior sample Metropolis step.

Given a trace t and stochastic function f depending on params..., generates proposals from prior draws and accepts based on the likelihood ratio.

CrimsonSkyline.mh_step — Method

function mh_step(t :: Trace, f; params = (), return_val :: Bool = false)

An independent prior sample Metropolis step.

Given a trace t and stochatic function f depending on params..., generates proposals from prior draws and accepts based on the likelihood ratio.

nested

CrimsonSkyline.nested — Method

function nested(f, replace_fn; params = (), num_points :: Int64 = 1)

Generic implementation of nested sampling (Skilling, Nested sampling for general Bayesian computation, Bayesian Analysis, 2006). The number of sampling iterations is a function of num_points aka $N$, and the empirical entropy of the sampling distribution, given at the $n$-th iteration by $H_n \approx \sum_k \hat p_k \log \hat p_k^{-1}$, where $\hat p_k = L_k w_k / Z_k$, $L_k$ is the likelihood value, $w_k$ is the difference in prior mass, and $Z_k$ is the current estimate of the partition function. The number of sampling iterations is equal to $\min_n \{n > 0: n > NH_n\}$.

Args:

f: stochastic function. Must have signature f(t :: Trace, params...)

replace_fn: function that returns a tuple (new_trace :: Trace, new_log_likelihood :: Float64). The input signature of this function must be replace_fn(f :: F, log_likelihood :: Float64, params...) where F <: Function. It must guarantee that new_log_likelihood > log_likelihood.

params: any parameters to pass to f

num_points: the number of likelihood points to keep track of

CrimsonSkyline.nested — Method

nested(f; params = (), num_points :: Int64 = 1)

Run nested sampling using internal rejection method.

results

CrimsonSkyline.NonparametricSamplingResults — Type

struct NonparametricSamplingResults{I} <: SamplingResults{I}
    interpretation :: I
    log_weights :: Array{Float64, 1}
    return_values :: Array{Any, 1}
    traces :: Array{Trace, 1}
end

Wrapper for results of sampling. Implements the following methods from Base: getindex, length, keys. Intepretation of log weights is dependent on I.

CrimsonSkyline.ParametricSamplingResults — Type

struct ParametricSamplingResults{I} <: SamplingResults{I}
    interpretation :: I
    log_weights :: Array{Float64, 1}
    return_values :: Array{Any, 1}
    traces :: Array{Trace, 1}
    distributions :: Dict
end

distributions maps from addresses to distributions, $a \mapsto \pi^{(a)}_\psi(z)$, where $\pi^{(a)}_\psi(z)$ solves

\[\max_\psi E_{z \sim p(z|x)}[\log \pi^{(a)}_\psi(z)].\]

The distributions are not used to generate values but only to score sampled values; values are still sampled from the posterior traces. Right now, the parametric approximation is very simple: values with support over the negative orthant of $\mathbb R^D$ are approximated by (multivariate) normal distributions, while values with support over only the positive orthant of $\mathbb{R}^D$ are approximated by (multivariate) lognormal distributions. This behavior is expected to change in the future.

CrimsonSkyline.addresses — Method

function addresses(r::SamplingResults{I}) where I <: InferenceType

Get all addresses associated with the SamplingResults object, $A = \bigcup_{t\in \text{traces}}\mathcal A_t$

CrimsonSkyline.aic — Method

function aic(r :: SamplingResults{I}) where I <: InferenceType

Computes an empirical estimate of the Akaike Information Criterion from a SamplingResults. The formula is

\[\text{AIC}(r)/2 = \min_{t \in \text{traces}(r)}|\text{params}(t)| - \hat\ell(t),\]

where $|\text{params}(t)|$ is the number of non-observed and non-deterministic sample nodes and $\hat\ell(t)$ is the empirical maximum likelihood.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, r::ParametricSamplingResults{I}; pa = ()) where I <: InferenceType

Treat a marginal site of a SamplingResults as a distribution, sampling from it into a trace.

CrimsonSkyline.sample — Method

function sample(t :: Trace, a, r :: NonparametricSamplingResults{I}; pa = ()) where I <: InferenceType

Treat a marginal site of a SamplingResults as a distribution, sampling from it into a trace.

CrimsonSkyline.to_parametric — Method

function to_parametric(r::NonparametricSamplingResults{I}) where I<:InferenceType

Converts a nonparametric sampling results object into one that additionally contains a mapping from addresses to distributions.

Distributions.logpdf — Method

function Distributions.logpdf(r :: A, v) where A <: AbstractArray

Interprets an array of objects as a delta distribution over those objects. If v is in the support set, returns $-log |r|$. Otherwise, returns $-\infty$.

online

CrimsonSkyline.combine — Method

function combine(v, pv::Vector{Float64})

Combines iterable of ParametricSamplingResults, v, into a single ParametricSamplingResults with mixture distribution parametric posteriors. pv is a weight vector (all elements > 0) that assigns relative weight to the mixture components in each ParametricSamplingResults. Mathematically, for each address $a$, the resulting mixture distribution is

\[p_a(z) = \sum_{r \in \mathtt{v}:\ a \in \mathcal{A}(r)} p^r_a(z)\ w^a_r,\]

where each $p^r_a(z)$ is the $r$-th ParametricSamplingResults's estimated parametric posterior marginal for address $a$, if it exists, and $w^a_r = \mathtt{pv}_r \Big/ \left( \sum_{r \in \mathtt{v}:\ a \in \mathcal{A}(r)} \mathtt{pv}_r\right)$.

CrimsonSkyline.reflate! — Method

function reflate!(r::ParametricSamplingResults{I}, mapping::Dict) where I<:InferenceType

Reflates the scale parameters of the distributions of r. The inferred scale parameters approximate the true uncertainty of the posterior distribution under the critical assumption that the data generating process (DGP) remains constant. When the DGP is nonstationary, it is necessary to intervene and increase the value of the scale parameters to properly account for uncertainty. mapping is a dict of address => value where value is either a float or a positive definite matrix depending on if the distribution associated with address is scalar- or vector-valued.

Index

CrimsonSkyline.NonparametricSamplingResults

CrimsonSkyline.ParametricNode

CrimsonSkyline.ParametricSamplingResults

CrimsonSkyline.SampleableNode

CrimsonSkyline.Trace

CrimsonSkyline.TypedTrace

CrimsonSkyline.UntypedTrace

CrimsonSkyline.accept

CrimsonSkyline.addresses

CrimsonSkyline.aic

CrimsonSkyline.aic

CrimsonSkyline.block

CrimsonSkyline.block

CrimsonSkyline.combine

CrimsonSkyline.condition

CrimsonSkyline.connect_pa_ch!

CrimsonSkyline.copy_common!

CrimsonSkyline.forward_sampling

CrimsonSkyline.importance_sampling

CrimsonSkyline.importance_sampling

CrimsonSkyline.input

CrimsonSkyline.interpret_latent!

CrimsonSkyline.is_step

CrimsonSkyline.is_step

CrimsonSkyline.likelihood_weighting

CrimsonSkyline.likelihood_weighting

CrimsonSkyline.likelihood_weighting_results

CrimsonSkyline.load

CrimsonSkyline.load_csm

CrimsonSkyline.load_jdb

CrimsonSkyline.log_acceptance_ratio

CrimsonSkyline.log_evidence

CrimsonSkyline.loglatent

CrimsonSkyline.loglikelihood

CrimsonSkyline.logprob

CrimsonSkyline.logprob

CrimsonSkyline.logprob!

CrimsonSkyline.lw_step

CrimsonSkyline.lw_step

CrimsonSkyline.mh

CrimsonSkyline.mh

CrimsonSkyline.mh

CrimsonSkyline.mh

CrimsonSkyline.mh_step

CrimsonSkyline.mh_step

CrimsonSkyline.mh_step

CrimsonSkyline.mh_step

CrimsonSkyline.nested

CrimsonSkyline.nested

CrimsonSkyline.node

CrimsonSkyline.node

CrimsonSkyline.node

CrimsonSkyline.node

CrimsonSkyline.normalized_weights

CrimsonSkyline.observe

CrimsonSkyline.pa_from_trace

CrimsonSkyline.plate

CrimsonSkyline.plate

CrimsonSkyline.plate

CrimsonSkyline.plate

CrimsonSkyline.plate

CrimsonSkyline.plate

CrimsonSkyline.propose

CrimsonSkyline.reflate!

CrimsonSkyline.rejection

CrimsonSkyline.replace

CrimsonSkyline.replace

CrimsonSkyline.replay

CrimsonSkyline.rewrite

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.sample

CrimsonSkyline.save

CrimsonSkyline.save

CrimsonSkyline.to_parametric

CrimsonSkyline.to_table

CrimsonSkyline.trace

CrimsonSkyline.trace

CrimsonSkyline.transform

CrimsonSkyline.update

Distributions.logpdf